# services/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AcademyAdminResponse:
    properties:
      academy:
        $ref: '#/definitions/dto.AcademyHoldingSimpleResponse'
      academy_id:
        type: string
      approved_at:
        type: string
      approved_by:
        $ref: '#/definitions/dto.UserSimpleResponse'
      approved_by_id:
        type: string
      branch:
        $ref: '#/definitions/dto.AcademyBranchSimpleResponse'
      branch_id:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserSimpleResponse'
      deleted_at:
        type: string
      deleted_by:
        $ref: '#/definitions/dto.UserSimpleResponse'
      id:
        type: string
      role:
        $ref: '#/definitions/dto.RoleSimpleResponse'
      role_id:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserSimpleResponse'
      user:
        $ref: '#/definitions/dto.UserSimpleResponse'
      user_id:
        type: string
    type: object
...
  dto.AcademyBranchSimpleResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      holding_id:
        type: string
      id:
        type: string
      member_count:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      sport_id:
        type: string
      sport_name:
        type: string
      status_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.AcademyHoldingSimpleResponse:
    properties:
      branches_count:
        type: integer
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      image_attachment_id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      status_id:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.AssignRolesRequest:
    properties:
      roles:
        items:
          type: string
        type: array
      user_id:
        example: 01950000-0000-7000-8000-000000000001
        type: string
    required:
    - roles
    - user_id
    type: object
...
  dto.DeleteUserRequest:
    properties:
      password:
        example: Password123!
        type: string
    required:
    - password
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: athlete@sportsentral.id
        type: string
    required:
    - email
    type: object
...
  dto.PermissionSimpleResponse:
    properties:
      action:
        type: string
      id:
        type: string
      resource:
        type: string
      slug:
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        example: athlete@sportsentral.id
        type: string
      full_name:
        example: John Doe
        type: string
      password:
        example: Password123!
        minLength: 8
        type: string
      role_id:
        example: 01950000-0000-7000-8000-000000000001
        type: string
      username:
        example: john_doe
        type: string
    required:
    - email
    - full_name
    - password
    - role_id
    - username
    type: object
...
  dto.RemoveRolesRequest:
    properties:
      roles:
        items:
          type: string
        type: array
      user_id:
        example: 01950000-0000-7000-8000-000000000001
        type: string
    required:
    - roles
    - user_id
    type: object
  dto.ResetPasswordRequest:
    properties:
      password:
        example: NewPassword123!
        minLength: 8
        type: string
      token:
        example: reset-token-uuid
        type: string
    required:
    - password
    - token
    type: object
...
  dto.RoleSimpleResponse:
    properties:
      id:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: string
        type: array
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionSimpleResponse'
        type: array
      slug:
        type: string
    type: object
...
  dto.SendVerifyEmailRequest:
    properties:
      email:
        example: athlete@sportsentral.id
        type: string
    required:
    - email
    type: object
...
  dto.StatusSimpleResponse:
    properties:
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      type:
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      full_name:
        example: John Doe Updated
        type: string
      username:
        example: john_updated
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      roles:
        items:
          $ref: '#/definitions/dto.UserRoleSimpleResponse'
        type: array
      status:
        $ref: '#/definitions/dto.StatusSimpleResponse'
      updated_at:
        type: string
      username:
        type: string
      verified_at:
        type: string
    type: object
  dto.UserRoleSimpleResponse:
    properties:
      expired_at:
        type: string
      role:
        $ref: '#/definitions/dto.RoleSimpleResponse'
      status:
        $ref: '#/definitions/dto.StatusSimpleResponse'
    type: object
  dto.UserSimpleResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
  dto.VerifyAccountRequest:
    properties:
      token:
        example: verification-token-uuid
        type: string
    required:
    - token
    type: object
...
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: dev@sportsentral.id
    name: Sport-Sentral Development Team
  description: Centralized API Documentation for Sport-Sentral Microservices Gateway
    (Auth, RBAC, Academy, Venue, Sport, Competition, Scout, Attachment, Log).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sport-Sentral API Gateway
  version: "1.0"
paths:
...
  /academy/admins/user/{userID}:
    get:
      parameters:
      - description: User ID (UUID)
        in: path
        name: userID
        required: true
        type: string
      - description: Holding ID (UUID)
        in: query
        name: holding_id
        type: string
      - description: Branch ID (UUID)
        in: query
        name: branch_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AcademyAdminResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get academy admin data by user ID
      tags:
      - Academies
...
  /forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Forgot password (Root alias)
      tags:
      - Users
...
  /me:
    delete:
      consumes:
      - application/json
      description: Soft-delete current user account with password confirmation.
      parameters:
      - description: Password Confirmation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete current user account
      tags:
      - Users
    get:
      description: Retrieve profile data, status, and active role of the currently
        authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get current logged-in user profile
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: Update username or full name of the currently logged-in account.
      parameters:
      - description: Profile Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
...
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user through the root registration endpoint.
      parameters:
      - description: New User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create new user (Direct register)
      tags:
      - Users
...
  /reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password Reset Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Reset password (Root alias)
      tags:
      - Users
...
  /users:
    get:
      description: Retrieve paginated list of all users (platform admin only).
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get list of users (Admin)
      tags:
      - Users
  /users/{id}:
    get:
      description: Retrieve detailed user profile information by user ID (UUID).
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
  /users/assign-roles:
    post:
      consumes:
      - application/json
      description: Assign a list of roles to a specific user account.
      parameters:
      - description: Assign Roles Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AssignRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Assign roles to user (Admin)
      tags:
      - Users
  /users/remove-roles:
    post:
      consumes:
      - application/json
      description: Remove a list of roles from a specific user account.
      parameters:
      - description: Remove Roles Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RemoveRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Remove roles from user (Admin)
      tags:
      - Users
...
  /verify-account:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verification Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Verify account (POST Body)
      tags:
      - Users
  /verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SendVerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send verification email (Root alias)
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in the format: `Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"